feat: check that date parameter are valid
