+++ package.json
@@ -47,6 +47,7 @@
     "@hapi/vision": "^7.0.0",
     "dotenv": "^16.3.1",
     "knex": "^2.5.1",
+    "moment": "^2.29.4",
     "lodash": "^4.17.21",
     "nodemon": "^3.0.1",
     "pg": "^8.11.1",
@@ -60,7 +61,6 @@
     "@types/express": "^4.17.17",
     "@types/hapi__hapi": "^21.0.0",
     "@types/mocha": "^10.0.1",
-    "@types/node": "^20.4.2",
     "@types/pino": "^7.0.5",
     "@types/pino-pretty": "^5.0.0",
     "@types/supertest": "^2.0.12",
